@@ -1,28 +1,12 @@
 <template>
   <a-config-provider :locale="locale">
-    <div id="app">
-      <router-view/>
-    </div>
+    <router-view />
   </a-config-provider>
 </template>
 
-<script>
-import { domTitle, setDocumentTitle } from '@/utils/domUtil'
-import { i18nRender } from '@/locales'
+<script setup lang="ts">
+import { ref } from 'vue'
+import zhCN from 'ant-design-vue/es/locale/zh_CN'
 
-export default {
-  data () {
-    return {
-    }
-  },
-  computed: {
-    locale () {
-      // 只是为了切换语言时，更新标题
-      const { title } = this.$route.meta
-      title && (setDocumentTitle(`${i18nRender(title)} - ${domTitle}`))
-
-      return this.$i18n.getLocaleMessage(this.$store.getters.lang).antLocale
-    }
-  }
-}
+const locale = ref(zhCN)
 </script>
