@@ -1,15 +1,17 @@
 <template>
-  <div>
-    AbsAdmin Home
-  </div>
+  <AdminLayout>
+    <div class="home-page">
+      <h1>欢迎使用 AbsAdmin 后台管理系统</h1>
+    </div>
+  </AdminLayout>
 </template>
 
-<script>
-export default {
-  name: 'home'
-}
+<script setup lang="ts">
+import AdminLayout from '@/components/AdminLayout.vue'
 </script>
 
 <style scoped>
-
+.home-page {
+  padding: 24px;
+}
 </style>
